+++ apps/fabric-website/src/components/Nav/Nav.tsx
@@ -99,11 +99,11 @@ export class Nav extends React.Component<INavProps, INavState> {
     const links = pages
       // Don't render pages that are explicitly told not to, as well as top-level pages that aren't active.
       .filter(page => (!page.hasOwnProperty('isHiddenFromMainNav') && !(page.isUhfLink && !_hasActiveChild(page))) || _isPageActive(page))
-      .map((page: INavPage, linkIndex: number) => {
+      .map((page: INavPage) => {
         if (page.isCategory && !filterState) {
-          return <>{page.pages.map((innerPage: INavPage, innerLinkIndex) => this._renderLink(innerPage, innerLinkIndex))}</>;
+          return <>{page.pages.map(innerPage => this._renderLink(innerPage))}</>;
         }
-        return page.isCategory && filterState ? this._renderCategory(page, linkIndex) : this._renderLink(page, linkIndex);
+        return page.isCategory && filterState ? this._renderCategory(page) : this._renderLink(page);
       });
 
     return (
@@ -116,16 +116,16 @@ export class Nav extends React.Component<INavProps, INavState> {
     );
   }
 
-  private _renderCategory(page: INavPage, categoryIndex: number): React.ReactElement<{}> {
+  private _renderCategory(page: INavPage): React.ReactElement<{}> {
     if (page.isCategory && page.pages) {
       return (
-        <li key={categoryIndex} className={css(styles.category, _hasActiveChild(page) && styles.hasActiveChild)}>
+        <li key={page.title + '-heading'} className={css(styles.category, _hasActiveChild(page) && styles.hasActiveChild)}>
           <CollapsibleSection
             title={{ text: page.title, styles: getTitleStyles }}
             styles={{ body: [{ marginLeft: '28px' }] }}
             defaultCollapsed={!_hasActiveChild(page)}
           >
-            <ul>{page.pages.map((innerPage: INavPage, indexNumber: number) => this._renderLink(innerPage, indexNumber))}</ul>
+            <ul>{page.pages.map(innerPage => this._renderLink(innerPage))}</ul>
           </CollapsibleSection>
         </li>
       );
@@ -147,7 +147,7 @@ export class Nav extends React.Component<INavProps, INavState> {
     return this._renderLinkList(links, true, title);
   }
 
-  private _renderLink(page: INavPage, linkIndex: number): React.ReactElement<{}> {
+  private _renderLink(page: INavPage): React.ReactElement<{}> {
     const ariaLabel = page.pages ? 'Hit enter to open sub menu, tab to access sub menu items.' : '';
     const title = page.title === 'Fabric' ? 'Home page' : page.title;
     const childLinks =
@@ -180,26 +180,24 @@ export class Nav extends React.Component<INavProps, INavState> {
     }
 
     return (
-      <>
-        <li
-          className={css(
-            styles.link,
-            _isPageActive(page) && searchQuery === '' ? styles.isActive : '',
-            _hasActiveChild(page) ? styles.hasActiveChild : '',
-            page.isHomePage ? styles.isHomePage : '',
-            page.className ? styles[page.className] : '',
-            page.isUhfLink ? styles.isUhfLink : ''
-          )}
-          key={linkIndex}
-        >
-          {!page.isUhfLink && (page.isFilterable && searchQuery !== '' ? matchIndex > -1 : true) && (
-            <a href={page.url} onClick={this._onLinkClick} title={title} aria-label={ariaLabel}>
-              {linkText}
-            </a>
-          )}
-          {childLinks}
-        </li>
-      </>
+      <li
+        className={css(
+          styles.link,
+          _isPageActive(page) && searchQuery === '' ? styles.isActive : '',
+          _hasActiveChild(page) ? styles.hasActiveChild : '',
+          page.isHomePage ? styles.isHomePage : '',
+          page.className ? styles[page.className] : '',
+          page.isUhfLink ? styles.isUhfLink : ''
+        )}
+        key={page.url || title}
+      >
+        {!page.isUhfLink && (page.isFilterable && searchQuery !== '' ? matchIndex > -1 : true) && (
+          <a href={page.url} onClick={this._onLinkClick} title={title} aria-label={ariaLabel}>
+            {linkText}
+          </a>
+        )}
+        {childLinks}
+      </li>
     );
   }