+++ apps/todo-app/src/components/TodoForm.tsx
@@ -47,9 +47,6 @@ export default class TodoForm extends BaseComponent<ITodoFormProps, ITodoFormSta
   constructor(props: ITodoFormProps) {
     super(props);
 
-    this._onSubmit = this._onSubmit.bind(this);
-    this._onBeforeTextFieldChange = this._onBeforeTextFieldChange.bind(this);
-
     this.state = {
       inputValue: '',
       errorMessage: ''
@@ -64,7 +61,7 @@ export default class TodoForm extends BaseComponent<ITodoFormProps, ITodoFormSta
           value={this.state.inputValue}
           componentRef={this._textField}
           placeholder={strings.inputBoxPlaceholder}
-          onBeforeChange={this._onBeforeTextFieldChange}
+          onChange={this._onTextFieldChange}
           autoComplete="off"
           errorMessage={this.state.errorMessage}
         />
@@ -98,12 +95,12 @@ export default class TodoForm extends BaseComponent<ITodoFormProps, ITodoFormSta
     }
   };
 
-  private _onBeforeTextFieldChange(newValue: string): void {
+  private _onTextFieldChange = (event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, newValue: string | undefined): void => {
     this.setState({
-      inputValue: newValue,
+      inputValue: newValue || '',
       errorMessage: ''
     });
-  }
+  };
 
   private _getTitleErrorMessage(title: string): string {
     if (title.trim() === '') {