+++ packages/dashboard/src/components/DetailPanel/Examples/DetailPanel.L2Content.Example.tsx
@@ -34,7 +34,7 @@ export class DetailPanelL2ContentExample extends React.PureComponent<{}, IDetail
         <DetailPanel
           mainHeader={header}
           mainContent={this.getMainContent()}
-          onDetailPanelDimiss={() => {
+          onDetailPanelDismiss={() => {
             this.setState({ show: false, currentL2Id: undefined });
           }}
           onL2BackClick={() => {