+++ packages/example-app-base/src/components/Header/Header.types.ts
@@ -1,0 +1,9 @@
+import { IWithResponsiveModeState } from 'office-ui-fabric-react/lib/utilities/decorators/withResponsiveMode';
+
+export interface IHeaderProps extends IWithResponsiveModeState {
+  title: string;
+  sideLinks: { name: string; url: string }[];
+
+  isMenuVisible: boolean;
+  onIsMenuVisibleChanged?: (isMenuVisible: boolean) => void;
+}