+++ packages/experiments/src/components/Button/examples/Button.Tokens.Example.tsx
@@ -5,11 +5,18 @@ import { createTheme, Spinner, Stack } from 'office-ui-fabric-react';
 const menuItems = [{ key: 'a', name: 'Item a' }, { key: 'b', name: 'Item b' }];
 const buttonMenu: IMenuButtonProps['menu'] = render => render((MenuType, props) => <MenuType {...props} items={menuItems} />);
 
-const sectionGap = 32;
+const tokens = {
+  sectionStack: {
+    childrenGap: 32
+  },
+  buttonStack: {
+    childrenGap: 8
+  }
+};
 
 const testTheme = createTheme({
   semanticColors: {
-    buttonText: 'red'
+    buttonText: '#E20000'
   },
   fonts: {
     medium: {
@@ -22,70 +29,70 @@ const testTheme = createTheme({
 export class ButtonTokensExample extends React.Component<{}, {}> {
   public render(): JSX.Element {
     const ButtonSet = (props: IButtonProps) => (
-      <Stack horizontal disableShrink verticalAlign="center" gap={8}>
+      <Stack horizontal disableShrink verticalAlign="center" tokens={tokens.buttonStack}>
         <Button {...props} />
         <Button {...props} primary />
         <Button {...props} disabled />
         <Button
           {...props}
           tokens={{
-            backgroundColor: 'red',
+            backgroundColor: '#E20000',
             backgroundColorHovered: 'pink',
             color: 'white',
-            colorHovered: 'white',
+            colorHovered: '#595959',
             iconColor: 'white',
-            iconColorHovered: 'white'
+            iconColorHovered: '#595959'
           }}
         />
       </Stack>
     );
 
     const MenuButtonSet = (props: IMenuButtonProps) => (
-      <Stack horizontal disableShrink verticalAlign="center" gap={8}>
+      <Stack horizontal disableShrink verticalAlign="center" tokens={tokens.buttonStack}>
         <MenuButton {...props} />
         <MenuButton {...props} primary />
         <MenuButton {...props} disabled />
         <MenuButton
           {...props}
           tokens={{
-            backgroundColor: 'red',
+            backgroundColor: '#E20000',
             backgroundColorHovered: 'pink',
             color: 'white',
-            colorHovered: 'white',
+            colorHovered: '#595959',
             iconColor: 'white',
-            iconColorHovered: 'white'
+            iconColorHovered: '#595959'
           }}
         />
       </Stack>
     );
 
     return (
-      <Stack gap={sectionGap}>
-        <ButtonSet />
+      <Stack tokens={tokens.sectionStack}>
+        <ButtonSet ariaLabel="Button with no icon or content" />
         <ButtonSet content="No Icon" />
-        <ButtonSet content={<Spinner />} />
+        <ButtonSet content={<Spinner />} ariaLabel="Button that has a Spinner as its content" />
         <ButtonSet icon="upload" content="Button with Icon" />
         <ButtonSet icon="upload" href="http://www.microsoft.com" content="Button with href" />
-        <ButtonSet circular />
+        <ButtonSet circular ariaLabel="Circular Button with no icon" />
         <ButtonSet circular icon="share" />
         <MenuButtonSet
           icon={render => render((IconType, iconProps) => <IconType {...iconProps} iconName="upload" />)}
           content="Menu button with icon"
           menu={buttonMenu}
         />
-        <Stack horizontal disableShrink verticalAlign="center" gap={8}>
+        <Stack horizontal disableShrink verticalAlign="center" tokens={tokens.buttonStack}>
           <Button
             primary
             icon="PeopleAdd"
             content="Token Function: Red BG, White Text"
             tokens={(props, theme) =>
               props.primary && {
-                backgroundColor: 'red',
+                backgroundColor: '#E20000',
                 backgroundColorHovered: 'pink',
                 color: 'white',
-                colorHovered: 'white',
+                colorHovered: '#595959',
                 iconColor: 'white',
-                iconColorHovered: 'white'
+                iconColorHovered: '#595959'
               }
             }
           />