+++ packages/office-ui-fabric-react/src/components/MarqueeSelection/examples/MarqueeSelection.Basic.Example.tsx
@@ -3,9 +3,8 @@ import * as React from 'react';
 import { css, createArray } from 'office-ui-fabric-react/lib/Utilities';
 import { Checkbox } from 'office-ui-fabric-react/lib/Checkbox';
 import { MarqueeSelection, Selection, IObjectWithKey } from 'office-ui-fabric-react/lib/MarqueeSelection';
-import './MarqueeSelection.Basic.Example.scss';
-import * as exampleStylesImport from '../../../common/_exampleStyles.scss';
-const exampleStyles: any = exampleStylesImport;
+import * as styles from './MarqueeSelection.Basic.Example.scss';
+import * as exampleStyles from '../../../common/_exampleStyles.scss';
 
 interface IPhoto extends IObjectWithKey {
   url: string;
@@ -59,11 +58,11 @@ export class MarqueeSelectionBasicExample extends React.Component<{}, IMarqueeSe
       <MarqueeSelection selection={this._selection} isEnabled={this.state.isMarqueeEnabled}>
         <Checkbox className={exampleStyles.exampleCheckbox} label="Is marquee enabled" defaultChecked={true} onChange={this._onChange} />
         <p>Drag a rectangle around the items below to select them:</p>
-        <ul className="ms-MarqueeSelectionBasicExample-photoList">
+        <ul className={styles.photoList}>
           {PHOTOS.map((photo, index) => (
             <div
               key={index}
-              className={css('ms-MarqueeSelectionBasicExample-photoCell', this._selection.isIndexSelected(index) && 'is-selected')}
+              className={css(styles.photoCell, this._selection.isIndexSelected(index) && 'is-selected')}
               data-is-focusable={true}
               data-selection-index={index}
               onClick={this._log('clicked')}