+++ packages/office-ui-fabric-react/src/components/Stack/Stack.types.ts
@@ -31,7 +31,7 @@ export interface IStackSlots {
 
 export interface IStackProps
   extends IStackSlots,
-    IStyleableComponentProps<IStackProps, IStackStyles, IStackTokens>,
+    IStyleableComponentProps<IStackProps, IStackTokens, IStackStyles>,
     React.HTMLAttributes<HTMLElement> {
   /**
    * Defines how to render the Stack.
@@ -110,6 +110,8 @@ export interface IStackProps
   wrap?: boolean;
 }
 
-export interface IStackTokens {}
+export interface IStackTokens {
+  childrenGap?: number | string;
+}
 
 export type IStackStyles = IComponentStyles<IStackSlots>;