+++ packages/office-ui-fabric-react/src/components/Stack/StackItem/StackItem.types.ts
@@ -45,8 +45,16 @@ export interface IStackItemProps extends IStackItemSlots, IStyleableComponentPro
    * @defaultvalue true
    */
   verticalFill?: boolean;
+
+  /**
+   * Defines order of the StackItem.
+   * @defaultvalue 0
+   */
+  order?: number | string;
 }
 
-export interface IStackItemTokens {}
+export interface IStackItemTokens {
+  margin?: number | string;
+}
 
 export type IStackItemStyles = IComponentStyles<IStackItemSlots>;