+++ packages/office-ui-fabric-react/src/components/Stack/examples/Stack.Horizontal.VerticalAlign.Example.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { Stack } from '../Stack';
+import { IStackTokens } from '../Stack.types';
 import { mergeStyleSets, DefaultPalette } from 'office-ui-fabric-react/lib/Styling';
 
 export class HorizontalStackVerticalAlignExample extends React.Component<{}, {}> {
@@ -22,8 +23,10 @@ export class HorizontalStackVerticalAlignExample extends React.Component<{}, {}>
       }
     });
 
+    const stackTokens: IStackTokens = { childrenGap: 5 };
+
     return (
-      <Stack gap={5}>
+      <Stack tokens={stackTokens}>
         <span>Top-aligned</span>
         <Stack horizontal verticalAlign="start" className={styles.root}>
           <span>1</span>