+++ packages/office-ui-fabric-react/src/components/pickers/BasePicker.styles.ts
@@ -78,7 +78,8 @@ export function getStyles(props: IBasePickerStyleProps): IBasePickerStyles {
       classNames.itemsWrapper,
       {
         display: 'flex',
-        flexWrap: 'wrap'
+        flexWrap: 'wrap',
+        maxWidth: '100%'
       }
     ],
     input: [