+++ packages/react-cards/CHANGELOG.md
@@ -1,6 +1,34 @@
 # Change Log - @uifabric/react-cards
 
-This log was last generated on Wed, 27 Mar 2019 12:34:02 GMT and should not be manually modified.
+This log was last generated on Fri, 05 Apr 2019 12:36:12 GMT and should not be manually modified.
+
+## 0.2.3
+Fri, 05 Apr 2019 12:36:12 GMT
+
+### Patches
+
+- Card: Supporting High Contrast mode.
+
+## 0.2.2
+Thu, 04 Apr 2019 12:34:57 GMT
+
+### Patches
+
+- Card: Updating examples to show narrow card with contents and to remove use of mergeStyleSets and gap.
+
+## 0.2.1
+Tue, 02 Apr 2019 00:38:15 GMT
+
+### Patches
+
+- Use ^ ranges instead of >=
+
+## 0.2.0
+Mon, 01 Apr 2019 12:37:03 GMT
+
+### Minor changes
+
+- Card: Improving styling and adding compact Card styling and examples.
 
 ## 0.1.2
 Wed, 27 Mar 2019 12:34:02 GMT